adding cloud functions, spinning off fetcher to a REST API app
